fetchCoinPrices triggered by fetchAccountBalances
